refactor: Rename HTTP repository to handler repository
